Fix: Focal loss implementation in training/loss.py
Why: Focal loss was not correctly implemented in training/loss.py,
leading to suboptimal training performance. This fix ensures that the
focal loss function is correctly applied, improving the model's ability
to focus on hard-to-classify examples and reducing the impact of easy
examples during training.

How: Originally, the focal loss alpha parameter was just a constant
while refering to the paper, alpha should be a tensor with the classes
weights. In the current implementation, you still have the option to
use a constant alpha but you can also pass a tensor with the
weights of each class.

Tags: Loss, FocalLoss, Training